# gmao/backend/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GMAO / MAINTX API",
    "description": "API REST pour la Gestion de Maintenance (ordres de travail, équipements, stock, effectif, rapports). Multi-tenant. Documentation détaillée intégration externe : docs/API_INTEGRATION_EXTERNE.md",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "/api", "description": "API relative" }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Connexion",
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "email": { "type": "string" }, "password": { "type": "string" } }, "required": ["email", "password"] } } } },
        "responses": { "200": { "description": "Token JWT" } }
      }
    },
    "/work-orders": {
      "get": { "summary": "Liste des OT", "parameters": [{ "name": "status", "in": "query", "schema": { "type": "string" } }, { "name": "page", "in": "query", "schema": { "type": "integer" } }], "responses": { "200": { "description": "Liste ou pagination" } } },
      "post": { "summary": "Créer un OT", "responses": { "201": { "description": "OT créé" } } }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    }
  },
  "security": [{ "bearerAuth": [] }]
}